requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
motor>=3.3.2
pymongo>=4.6.0
python-dotenv>=1.0.0
bcryptjs>=0.2.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pydantic>=2.5.0
email-validator>=2.1.0
httpx>=0.25.2
pytest>=7.4.3
pytest-asyncio>=0.21.1
aiofiles>=23.2.1
Pillow>=10.1.0
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.11.0
google-generativeai>=0.3.2
requests>=2.31.0